Fixed bug for number of remaining items when multiple games played
There was a bug where the number of items remaining was being carrier
over when more than one game was played in a row.  The map was being
modified during game 1 but not refreshed for game 2.  I have added code
that refreshes the US States and US State Capitals objects when a new
game is started or a game is re-tried.