Reformatted user prompts in gameplay page
Success/fail messages are now displayed above the prompt for the current/
next question.

Got rid of the fade effect. Now messages persist until they are replaced.
This eliminated the problem where guesses in close succession would
interfere with the fading and showing of the status message element.

Also refactored code to eliminate the need for the game-play-feedback-detail
span.  Everything now goes in game-play-feedback.