Updated game play feedback with more detail on incorrect guess
Previously, the game play feedback simply said "incorrect" for an
incorrect guess.  The problem was that for multiple consecutive incorrect
guesses, the status message didn't change so it wasn't clear to the user
that anything was happening (other than the wrong guess count incrementing
which wasn't obvious).  Now, assuming the user guesses a different wrong
state each time, they will see a different message, making it clear that
the app is registering each click.