Fixed bug impacting calculation of percent complete for won games
I noticed that games I had won were showing up as 98% complete.  Upon
investigation, this is because the map_remaining array for completed games
contained 1 value, an empty text string. As a result, the percent complete
calculation was counting this as 1 incomplete item, hence a result of
slightly lower than 100%.  To fix this, in situations where the
game_result === 'Won' I've overwritten percentComplete to be 100.